+++ buckler/software/apps/final_project/controller.h
@@ -13,6 +13,6 @@ typedef enum {
   PENDING,
 } robot_state_t;
 
-static double get_distance(uint16_t current_encoder, uint16_t prev_encoder);
+static float get_distance(uint16_t current_encoder, uint16_t prev_encoder);
 
 robot_state_t controller(robot_state_t state);
